requirements.txt
Django==4.1.2
Faker==	15.0.0
PyYAML==6.0
asgiref	==3.5.2
attrs==	22.1.0
config==0.5.1
django-debug-toolbar==3.7.0
django-filter==	22.1
djangorestframework==3.14.0
drf-spectacular==0.24.2
inflection==0.5.1
jsonschema==4.16.0
pip==22.2.2
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.4
setuptools==60.2.0
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
wheel==0.37.1